버그수정
- [x]  진화가이드창 위치 변경
- [x]  가공소 석탄 소모량 다시 체크
- [x]  금광,가공소 레벨이 0일때는 채굴하지 않기
- [x]  상점에 누락된 이미지 추가
- [x]  유니온 강화 사운드 및 레벨 증가 표시
- [x]  오버플로우 체크하기
- [x]  튜토리얼 진행중 UI 위치 수정
- [x]  배틀패스 30스테이지 슬롯 활성화 후 리워드 버튼 어둡게 보임

+++ Assets/WOM/Scripts/CustomPopup/PopupController.cs
@@ -137,7 +137,7 @@ public class PopupController : MonoBehaviour
 
         for (int i = 0; i < rewards.Count; i++)
         {
-            GlobalData.instance.rewardManager.RewardByType(rewards[i].type, (int)rewards[i].amount);
+            GlobalData.instance.rewardManager.RewardByType(rewards[i].type, rewards[i].amount);
         }
 
         //rewardEffect.SetActive(false);
